* Patched modifications made in opentibia SVN [4230, 4224] Fixes should decrease a number of crashes
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1797 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ creature.h
@@ -398,7 +398,10 @@ class Creature : public AutoID, virtual public Thing
 		bool isInternalRemoved;
 		bool isMapLoaded;
 		bool isUpdatingPath;
-		size_t checkCreatureVectorIndex;
+		// The creature onThink event vector this creature belongs to
+		// -1 represents that the creature isn't in any vector
+		// 0 represents it's going to be added in the next checkCreature call
+		int32_t checkCreatureVectorIndex;
 		int32_t health, healthMax;
 		int32_t mana, manaMax;
